Move clear command to person group
Renamed add delivery window to modify delivery window
