+++ src/test/java/seedu/address/logic/commands/ClearCommandTest.java
@@ -5,6 +5,7 @@ import static seedu.address.testutil.TypicalPersons.getTypicalAddressBook;
 
 import org.junit.jupiter.api.Test;
 
+import seedu.address.logic.commands.person.ClearCommand;
 import seedu.address.model.AddressBook;
 import seedu.address.model.DeliveryJobSystem;
 import seedu.address.model.Model;
